Enable writing plugins for jinja2 tests
