+++ v2/ansible/plugins/__init__.py
@@ -311,6 +311,13 @@ filter_loader = PluginLoader(
     'filter_plugins'
 )
 
+test_loader = PluginLoader(
+    'TestModule',
+    'ansible.plugins.test',
+    C.DEFAULT_TEST_PLUGIN_PATH,
+    'test_plugins'
+)
+
 fragment_loader = PluginLoader(
     'ModuleDocFragment',
     'ansible.utils.module_docs_fragments',
