+++ v2/ansible/template/__init__.py
@@ -28,7 +28,7 @@ from jinja2.runtime import StrictUndefined
 
 from ansible import constants as C
 from ansible.errors import AnsibleError, AnsibleFilterError, AnsibleUndefinedVariable
-from ansible.plugins import filter_loader, lookup_loader
+from ansible.plugins import filter_loader, lookup_loader, test_loader
 from ansible.template.safe_eval import safe_eval
 from ansible.template.template import AnsibleJ2Template
 from ansible.template.vars import AnsibleJ2Vars
@@ -57,6 +57,7 @@ class Templar:
         self._loader              = loader
         self._basedir             = loader.get_basedir()
         self._filters             = None
+        self._tests               = None
         self._available_variables = variables
 
         # flags to determine whether certain failures during templating
@@ -93,11 +94,28 @@ class Templar:
         self._filters = dict()
         for fp in plugins:
             self._filters.update(fp.filters())
+        self._filters.update(self._get_tests())
 
         return self._filters.copy()
 
+    def _get_tests(self):
+        '''
+        Returns tests plugins, after loading and caching them if need be
+        '''
+
+        if self._tests is not None:
+            return self._tests.copy()
+
+        plugins = [x for x in test_loader.all()]
+
+        self._tests = dict()
+        for fp in plugins:
+            self._tests.update(fp.tests())
+
+        return self._tests.copy()
+
     def _get_extensions(self):
-        ''' 
+        '''
         Return jinja2 extensions to load.
 
         If some extensions are set via jinja_extensions in ansible.cfg, we try
@@ -229,6 +247,7 @@ class Templar:
 
             environment = Environment(trim_blocks=True, undefined=StrictUndefined, extensions=self._get_extensions(), finalize=self._finalize)
             environment.filters.update(self._get_filters())
+            environment.tests.update(self._get_tests())
             environment.template_class = AnsibleJ2Template
 
             # FIXME: may not be required anymore, as the basedir stuff will
